CHANGE: renamed Account.ts => AccountModel.ts
