+++ src/services/AccountService.ts
@@ -1,6 +1,6 @@
 import { Inject, Service } from "@tsed/di";
 import { MongooseModel } from "@tsed/mongoose";
-import { AccountModel } from "../models/Account";
+import { AccountModel } from "../models/AccountModel";
 import { CRUDServiceBase } from "./CRUDServiceBase`1";
 
 @Service()
